Add block attributes
This adds the possibility to add block level attributes:

    {#id2}
    > line

When enabled to will generate (for HTML) an id="id2" in the openening
tag of the block quote:

    <blockquote id="id2">
    <p>line</p>
    </blockquote>

Only one block level attribute is allowed, if multiple are given only
the last one seen is applied, the rest is silently ignored.

The attributes are stored in *Parser and copied to the node in addBlock.
It's up to the renderer to use the attributes, the html renderer
currently applies them to most block level elements.

Signed-off-by: Miek Gieben <miek@miek.nl>