Add ParserFunc
Add a way to add custom parser functions that are called from the main
parser. This PR only adds a block level parsing option. It also requires
to make public: p.addChild, p.block and p.finalize.

With this in place it should be possible to parse a TOML title block
which will then be picked up the */renderer.

The test function in option_test.go is heavily inspired by function
`p.quote(data []byte) int`, (which does call addBlock and finalize).

We could potentially *always* call p.finalize and friends, but I'm not
sure if that make sense and in the case of a toml title block I don't
think I need/want that.

This avoids having a dependency on the toml Go library in this package.

Signed-off-by: Miek Gieben <miek@miek.nl>