parser: Implement Subfigures
This adds new block syntax that allows you to put multiple figures (or
other block elements) in a larger figure; essentially having subfigures.
Is only enabled when the Mmark extension is on.

The trigger is `!---` as it looks like a code block. This also allows
you to set block level attributes on the element.

See https://mmark.nl/post/syntax/#figures-and-subfigures for some extra
wording on the new syntax.

Signed-off-by: Miek Gieben <miek@miek.nl>