Fix definition list in definition lists
Because we a line by line parser with are too late to detect nested
definition lists. The fix implemented here detects that we are in this
situation and then backtracks to search for the beginning of the
previous line and starts the sublist there.

This doesn't break any existing test. Added def-in-def tests for this
particular case.

Signed-off-by: Miek Gieben <miek@miek.nl>