links: keep deferred status
Keep track if a link or footnote was deferred. This makes it possible to
do this again when generating (formatted) markdown output. I.e. now all
deferred links/footnotes become inline because it does not know that
some of them had been deferred.

Without this patch:

    Footnote^[footnote *with* stuff ]

With this patch:

    Footnote[^id]

    [^id]: footnote *with* stuff

Where the later is how to the original markdown file was formatted.

Signed-off-by: Miek Gieben <miek@miek.nl>