Set delimiter in ast.List as well.
Turns out when using the just merged ordered list stuff for allowing )
as a delimiter can't be picked up by clients because the ast.List
doesn't have it set. Make the oli detection in block.go check for the
delimiter as well and extend 'p.list()' with a delimiter parameter.

Note that p.list() now has 2 parameters solely for ordered lists, which
is a bit excessive. Tested this code with mmark to see if I can actually
use the detected delim to steer my output (that works).

Didn't add a test because we don't really have the unit test setup to
test internals like this.

Signed-off-by: Miek Gieben <miek@miek.nl>