Make tables break a paragraph
Tables are block level entities and as such should also break a
paragraph, just like lists, codeblocks, etc.

Add the missing the check in the paragraph() function. Needed to alter
the tableHeader() function to not added a TableHeader if detected, added
doRender bool for that.

testdata/Tables.tests extended with a test that has text before the
table and double checked it failed without the new code.

Fixes: #214

Signed-off-by: Miek Gieben <miek@miek.nl>