Allow suffix for cross references
See issue #223 for initial discussion. This is implemented in this PR,
added tests in ref_test.go.

Only issue is that: `(#y,es)` is now a valid reference, with the ID
being set to "y" - which may be surprising? Potentially this can be
hidden behind (another) extension?

The XMLv3 standard allows a 'format' attribute, so this is not going
away and provides a neat way to be able to set that attribute.

Signed-off-by: Miek Gieben <miek@miek.nl>