Add SVG element
SVG is now part of HTML 5 and the SVG element a valid top-level
element. The other SVG element, however, do not stand on their own.
https://html.spec.whatwg.org/dev/embedded-content-other.html#svg-0

Note: This does not match up with the definition of CommonMark HTML
blocks. CommonMark does not recognize the SVG element as a HTML block.
In this respect, however, it would seem that CommonMark is lagging
behind. https://spec.commonmark.org/0.31.2/#html-blocks